requirements.txt
openai>=1.0.0
faiss-cpu
flask
numpy
tqdm